pyproject.toml
[project]
name = "memory-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "fastmcp>=2.3.4",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.18",
    "openai>=1.79.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "pytest-dotenv>=0.5.2",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "qdrant-client>=1.14.2",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.34.2",
]
